+++ pkgs/native_assets_builder/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.17.0-wip
+
+- Bump `package:native_assets_cli` to 0.17.0.
+
 ## 0.16.0
 
 - Pass in path to pubspec (to read user-defines) in this package rather than in