+++ pkgs/native_toolchain_c/CHANGELOG.md
@@ -1,3 +1,7 @@
+## 0.14.0-wip
+
+- Bump `package:native_assets_cli` to 0.17.0.
+
 ## 0.13.0
 
 - Bump `package:native_assets_cli` to 0.16.0.